Error handling for entity

+++ lwjgl3/src/main/java/io/github/some_example_name/lwjgl3/Bucket.java
@@ -10,6 +10,7 @@ import com.badlogic.gdx.math.Rectangle;
  * so the CollisionManager can detect bucket-droplet collisions.
  * On collision, plays a click sound via IAudioSystem (DIP).
  *
+ * Change: width and height parameters are now float to match TextureObject.
  */
 public class Bucket extends TextureObject implements Collidable {
 
@@ -18,7 +19,18 @@ public class Bucket extends TextureObject implements Collidable {
     private final float speed;
     private final Rectangle bounds;
 
-    public Bucket(Texture texture, float x, float y, int width, int height,
+    /**
+     * Create a new bucket entity.
+     *
+     * @param texture     Texture to render. Must not be {@code null}.
+     * @param x           Initial X position. Must be finite.
+     * @param y           Initial Y position. Must be finite.
+     * @param width       Width in pixels. Must be finite and positive.
+     * @param height      Height in pixels. Must be finite and positive.
+     * @param audioSystem Audio system for collision sounds. May be {@code null}.
+     * @throws IllegalArgumentException if texture is null or dimensions are invalid.
+     */
+    public Bucket(Texture texture, float x, float y, float width, float height,
                   IAudioSystem audioSystem) {
         super(texture, x, y, height, width);
         this.audioSystem = audioSystem;
